Adding Last Year Columns
This script take an input dataset and add a column for the value of last year but adding missing accounts that can exist in the past or in the future.

The total sum of new Last Year column must be the same as sum of all values of last year dates.